Enable query commands on validation view models
