+++ src/MVVMFluent/ValidationViewModelBase.cs
@@ -1,7 +1,10 @@
+using MVVMFluent.Commands;
 using MVVMFluent.Interfaces;
+using MVVMFluent.Queries;
 using MVVMFluent.Validation;
 using System;
 using System.Collections;
+using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Linq;
@@ -14,6 +17,9 @@ namespace MVVMFluent;
 /// </summary>
 public abstract class ValidationViewModelBase : FluentSetterViewModelBase, IValidationFluentSetterViewModel
 {
+    private readonly Dictionary<string, object> _queryBuilderStore = new();
+    private IQueryDispatcher? _queryDispatcher;
+
     /// <summary>
     /// Gets a value indicating whether any properties in this view model have validation errors.
     /// </summary>
@@ -36,6 +42,16 @@ public abstract class ValidationViewModelBase : FluentSetterViewModelBase, IVali
     protected ValidationViewModelBase()
     {
         ErrorsChanged += ErrorsChangedHandler;
+    }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="ValidationViewModelBase"/> class with the specified query dispatcher.
+    /// </summary>
+    /// <param name="queryDispatcher">The dispatcher responsible for executing queries.</param>
+    protected ValidationViewModelBase(IQueryDispatcher queryDispatcher)
+        : this()
+    {
+        UseQueries(queryDispatcher);
     }
 
     private void ErrorsChangedHandler(object? sender, DataErrorsChangedEventArgs e)
@@ -146,6 +162,50 @@ public abstract class ValidationViewModelBase : FluentSetterViewModelBase, IVali
         return new ValidationFluentSetterBuilder<TValue>(value, this, propertyName);
     }
 
+    /// <summary>
+    /// Configures the view model to use the supplied <see cref="IQueryDispatcher"/>.
+    /// </summary>
+    /// <param name="queryDispatcher">The dispatcher responsible for executing queries.</param>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="queryDispatcher"/> is <see langword="null"/>.</exception>
+    protected void UseQueries(IQueryDispatcher queryDispatcher)
+    {
+        _queryDispatcher = queryDispatcher ?? throw new ArgumentNullException(nameof(queryDispatcher));
+    }
+
+    /// <summary>
+    /// Creates a query command builder that dispatches the supplied query when executed.
+    /// </summary>
+    /// <typeparam name="TResult">The type of result produced by the query.</typeparam>
+    /// <param name="factory">The factory used to create the query instance.</param>
+    /// <param name="propertyName">The property name used for caching. Automatically provided by the compiler.</param>
+    /// <returns>A <see cref="QueryAsyncCommandBuilder{TResult}"/> for configuring the command.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="factory"/> or <paramref name="propertyName"/> is null.</exception>
+    /// <exception cref="InvalidOperationException">Thrown when no query dispatcher has been configured.</exception>
+    protected QueryAsyncCommandBuilder<TResult> Send<TResult>(Func<IQuery<TResult>> factory, [CallerMemberName] string? propertyName = null)
+    {
+        if (factory == null)
+        {
+            throw new ArgumentNullException(nameof(factory));
+        }
+
+        EnsurePropertyName(propertyName);
+
+        if (_queryDispatcher == null)
+        {
+            throw new InvalidOperationException("No IQueryDispatcher has been configured. Provide one via the constructor or call UseQueries before creating query commands.");
+        }
+
+        if (_queryBuilderStore.TryGetValue(propertyName!, out var existingBuilder))
+        {
+            return (QueryAsyncCommandBuilder<TResult>)existingBuilder;
+        }
+
+        var builder = new QueryAsyncCommandBuilder<TResult>(this, factory, _queryDispatcher);
+        _queryBuilderStore[propertyName!] = builder;
+        _commandStore[propertyName!] = builder.Command;
+        return builder;
+    }
+
     private IEnumerable GetAllErrors()
     {
         return _builderStore.Values
@@ -158,9 +218,23 @@ public abstract class ValidationViewModelBase : FluentSetterViewModelBase, IVali
     /// </summary>
     protected override void DisposeInternal()
     {
+        _queryBuilderStore.Clear();
         ErrorsChanged -= ErrorsChangedHandler;
         ErrorsChanged = null;
 
         base.DisposeInternal();
+    }
+
+    private static void EnsurePropertyName(string? propertyName)
+    {
+        if (propertyName == null)
+        {
+            throw new ArgumentNullException(nameof(propertyName), "Not able to determine property name.");
+        }
+
+        if (propertyName == ".ctor")
+        {
+            throw new ArgumentException("Property name must be provided when it is used inside a constructor.", nameof(propertyName));
+        }
     }
 }
